Improve healthchecks.io pings
Send the healthchecks.io /start ping in the first task that runs, and send the final ping at the end of script execution only if a /start ping was sent. This allows us to measure the total run time of all due tasks, not just for back up, and prevents triggering an ‘up’ alert followed by a ‘down’ alert in the case when back up would finish successfully, but then some other part of the script would fail.